EventBuffer.Version nad AggregateVersion as long (#39)

+++ src/Qowaiv.DomainModel.TestTools/AggregateRootAssert.cs
@@ -100,7 +100,7 @@ public static class AggregateRootAssert
     }
 
     [Impure]
-    private static bool AppendEvents(this StringBuilder sb, int index, object exp, object act)
+    private static bool AppendEvents(this StringBuilder sb, long index, object exp, object act)
     {
         if (sb.AppendDifferentTypes(index, exp, act))
         {
@@ -116,7 +116,7 @@ public static class AggregateRootAssert
     }
 
     [Impure]
-    private static bool AppendDifferentTypes(this StringBuilder sb, int index, object exp, object act)
+    private static bool AppendDifferentTypes(this StringBuilder sb, long index, object exp, object act)
     {
         var actType = act.GetType();
         var expType = exp.GetType();
@@ -130,7 +130,7 @@ public static class AggregateRootAssert
     }
 
     [Impure]
-    private static bool AppendDifferentEvents(this StringBuilder sb, int index, object exp, object act)
+    private static bool AppendDifferentEvents(this StringBuilder sb, long index, object exp, object act)
     {
         var failure = false;
 
@@ -189,14 +189,14 @@ public static class AggregateRootAssert
     }
 
     [Impure]
-    private static bool AppendIdenticalEvents(this StringBuilder sb, int index, object @event)
+    private static bool AppendIdenticalEvents(this StringBuilder sb, long index, object @event)
     {
         sb.AppendLine($"[{index}] {@event.GetType().Name}");
         return false;
     }
 
     [Impure]
-    private static bool AppendExtraEvents(this StringBuilder sb, IEnumerable<object> events, int offset, int skip, string prefix)
+    private static bool AppendExtraEvents(this StringBuilder sb, IEnumerable<object> events, long offset, int skip, string prefix)
     {
         var index = offset + skip;
 
@@ -213,7 +213,7 @@ public static class AggregateRootAssert
     }
 
     [Impure]
-    private static bool AppendExpectedActual(this StringBuilder sb, int index, object expected, object actual)
+    private static bool AppendExpectedActual(this StringBuilder sb, long index, object expected, object actual)
     {
         var prefix = $"[{index}] ";
         var empty = new string(' ', prefix.Length);
