+++ test/Qowaiv.DomainModel.UnitTests/Models/Events/StoredEvent.cs
@@ -1,3 +1,3 @@
 ﻿namespace Models.Events;
 
-public sealed record StoredEvent(object Id, int Version, object Payload);
+public sealed record StoredEvent(object Id, long Version, object Payload);
